Classify ingests according to dev status
