+++ monitoring/daily_reporter/requirements.in
@@ -1,2 +1,3 @@
 boto3
 httpx
+jinja2
