+++ monitoring/daily_reporter/src/daily_reporter.py
@@ -1,107 +1,9 @@
 #!/usr/bin/env python
 
-import datetime
-import json
+import collections
 
-import boto3
-import httpx
-
-from elasticsearch import get_es_client
-
-
-def get_interesting_ingests(es_client, *, index_name, days_to_fetch):
-    """
-    Get ingests that are "interesting".
-    """
-    today = datetime.datetime.now()
-
-    start = today - datetime.timedelta(days=days_to_fetch)
-
-    # Did we find everything we were looking for?  It's possible we might
-    found_everything = True
-
-    # Ingests we want to return
-    ingests = {}
-
-    # We make three requests to Elasticsearch, to maximise the chance of finding
-    # ingests that might need more attention:
-    #
-    #   * everything modified in the last N days
-    #   * everything modified in the last N days which hasn't succeeded
-    #   * everything which hasn't completed
-    #
-    # A single ES request returns up to 10000 documents; if more than that have
-    # been updated (e.g. we've just done a big migration), we want to flag that
-    # fact and try to find the most interesting stuff.
-    filters = [
-        [{"range": {"lastModifiedDate": {"gte": start.strftime("%Y-%m-%d")}}}],
-        [{
-            "range": {"lastModifiedDate": {"gte": start.strftime("%Y-%m-%d")}}},{
-            "terms": {"status.id": ["processing", "accepted", "failed"]},
-        }],
-        [{"terms": {"status.id": ["processing", "accepted"]}}],
-    ]
-
-    for filter_clause in filters:
-        resp = es_client.request(
-            method="GET",
-            url=f"/{index_name}/_search",
-            json={
-                "query": {"bool": {"filter": filter_clause}},
-                "_source": [
-                    "id",
-                    "lastModifiedDate",
-                    "space.id",
-                    "status.id",
-                    "bag.info.externalIdentifier",
-                    "bag.info.version",
-                    "createdDate",
-                    "events.description",
-                ],
-                # We can retrieve up to 10000 documents in one request.
-                "size": 10000,
-                "sort": [{"id": "desc"}],
-            },
-        )
-
-        # We use a dict to de-duplicate result.  The same ingest may appear
-        # multiple times in this response -- that's not an issue.
-        for hit in resp.json()["hits"]["hits"]:
-            ingests[hit["_id"]] = _parse_ingest_from_hit(hit)
-
-        # If there are more than 10000 results, that suggests there are more
-        # that we didn't get.  Flag this to the caller.
-        if resp.json()["hits"]["total"]["value"] >= 10000:
-            found_everything = False
-
-    return {"ingests": ingests, "found_everything": found_everything}
-
-
-def _parse_date(d):
-    try:
-        return datetime.datetime.strptime(d, "%Y-%m-%dT%H:%M:%S.%fZ")
-    except ValueError:
-        return datetime.datetime.strptime(d, "%Y-%m-%dT%H:%M:%SZ")
-
-
-def _parse_ingest_from_hit(hit):
-    source = hit["_source"]
-    ingest = {
-        "id": source["id"],
-        "space": source["space"]["id"],
-        "status": source["status"]["id"],
-        "externalIdentifier": source["bag"]["info"]["externalIdentifier"],
-        "version": source["bag"].get("version"),
-        "createdDate": _parse_date(source["lastModifiedDate"]),
-        "events": [ev["description"] for ev in source["events"]]
-    }
-
-    try:
-        ingest["lastModifiedDate"] = _parse_date(source["lastModifiedDate"])
-    except KeyError:
-        ingest["lastModifiedDate"] = ingest["createdDate"]
-
-    return ingest
+from elasticsearch import get_es_client, get_interesting_ingests
+from ingests import classify_ingest
 
 
 def get_ingests_by_status(recent_ingests):
@@ -233,13 +135,18 @@ def main(*args):
     prod_ingests = get_interesting_ingests(
         es_client, index_name="storage_ingests", days_to_fetch=2
     )
-    # stage_ingests = get_recent_ingests(
-    #     es_client, index_name="storage_stage_ingests", days_to_fetch=2
-    # )
+
+    classified_ingests = {
+        'prod': collections.defaultdict(list),
+        'stage': collections.defaultdict(list),
+    }
+    for ingest in prod_ingests["ingests"]:
+        status = classify_ingest(ingest)
+        classified_ingests['prod'][status].append(ingest)
 
     from pprint import pprint
 
-    pprint(prod_ingests["ingests"])
+    pprint(classified_ingests['prod'].keys())
 
     #
     #
